Move shared service functions from websocket.py to helpers.py
