+++ custom_components/lock_code_manager/websocket.py
@@ -98,6 +98,12 @@ from .const import (
     EXCLUDED_CONDITION_PLATFORMS,
 )
 from .data import get_entry_data
+from .helpers import (
+    async_clear_slot_condition,
+    async_clear_usercode,
+    async_set_slot_condition,
+    async_set_usercode,
+)
 from .models import SlotCode, SlotEntityData, SlotEntityIds, SlotMetadata
 from .providers import BaseLock
 
@@ -1163,105 +1169,6 @@ def _get_lock_or_error(
         )
         return None
     return lock
-
-
-# =============================================================================
-# Shared Logic Functions (used by both websocket handlers and services)
-# =============================================================================
-
-
-def _get_managed_lock(hass: HomeAssistant, lock_entity_id: str) -> BaseLock:
-    """Get a managed lock by entity ID, raising if not found."""
-    lock = hass.data.get(DOMAIN, {}).get(CONF_LOCKS, {}).get(lock_entity_id)
-    if not lock:
-        raise ServiceValidationError(
-            f"Lock {lock_entity_id} is not managed by Lock Code Manager"
-        )
-    return lock
-
-
-async def async_set_usercode(
-    hass: HomeAssistant, lock_entity_id: str, code_slot: int, usercode: str
-) -> None:
-    """Set a usercode on a lock slot."""
-    lock = _get_managed_lock(hass, lock_entity_id)
-    await lock.async_internal_set_usercode(code_slot, usercode)
-
-
-async def async_clear_usercode(
-    hass: HomeAssistant, lock_entity_id: str, code_slot: int
-) -> None:
-    """Clear a usercode from a lock slot."""
-    lock = _get_managed_lock(hass, lock_entity_id)
-    await lock.async_internal_clear_usercode(code_slot)
-
-
-def _get_slot_config(config_entry: ConfigEntry, slot_num: int) -> dict[str, Any]:
-    """Get slot config dict, raising if not found."""
-    slots = get_entry_data(config_entry, CONF_SLOTS, {})
-    slot_key = slot_num if slot_num in slots else str(slot_num)
-    if slot_key not in slots:
-        raise ServiceValidationError(f"Slot {slot_num} not found in config entry")
-    return slots[slot_key]
-
-
-def _get_loaded_config_entry(hass: HomeAssistant, config_entry_id: str) -> ConfigEntry:
-    """Get a loaded config entry by ID, raising if not found or not loaded."""
-    config_entry = hass.config_entries.async_get_entry(config_entry_id)
-    if not config_entry:
-        raise ServiceValidationError(
-            f"No lock code manager config entry with ID `{config_entry_id}` found"
-        )
-    if config_entry.state is not ConfigEntryState.LOADED:
-        raise ServiceValidationError(f"Config entry {config_entry.entry_id} not loaded")
-    return config_entry
-
-
-async def async_set_slot_condition(
-    hass: HomeAssistant, config_entry_id: str, slot: int, entity_id: str
-) -> None:
-    """Set a condition entity for a slot."""
-    config_entry = _get_loaded_config_entry(hass, config_entry_id)
-    _get_slot_config(config_entry, slot)
-
-    # Verify entity exists
-    if not hass.states.get(entity_id):
-        raise ServiceValidationError(f"Entity {entity_id} not found")
-
-    # Check for excluded platforms
-    ent_reg = er.async_get(hass)
-    entity_entry = ent_reg.async_get(entity_id)
-    if entity_entry and entity_entry.platform in EXCLUDED_CONDITION_PLATFORMS:
-        raise ServiceValidationError(
-            f"Entities from the '{entity_entry.platform}' integration are not "
-            "supported as condition entities. See the wiki for details: "
-            "https://github.com/raman325/lock_code_manager/wiki/"
-            "Unsupported-Condition-Entity-Integrations"
-        )
-
-    # Update config entry data
-    data = copy.deepcopy(dict(config_entry.data))
-    slot_key = slot if slot in data[CONF_SLOTS] else str(slot)
-    data[CONF_SLOTS][slot_key][CONF_ENTITY_ID] = entity_id
-
-    if data != config_entry.data:
-        hass.config_entries.async_update_entry(config_entry, options=data)
-
-
-async def async_clear_slot_condition(
-    hass: HomeAssistant, config_entry_id: str, slot: int
-) -> None:
-    """Clear the condition entity from a slot."""
-    config_entry = _get_loaded_config_entry(hass, config_entry_id)
-    _get_slot_config(config_entry, slot)
-
-    # Update config entry data
-    data = copy.deepcopy(dict(config_entry.data))
-    slot_key = slot if slot in data[CONF_SLOTS] else str(slot)
-    data[CONF_SLOTS][slot_key].pop(CONF_ENTITY_ID, None)
-
-    if data != config_entry.data:
-        hass.config_entries.async_update_entry(config_entry, options=data)
 
 
 @websocket_api.websocket_command(
